Function is not copied, but imported into script. Now, `clone_data.py` depends on `ocean_emulators`.
