+++ src/ocean_emulators/utils/data.py
@@ -1,6 +1,7 @@
 import dataclasses
 import logging
 import re
+from collections import defaultdict
 from collections.abc import Callable
 from functools import cached_property
 from typing import Literal, Self
@@ -710,3 +711,29 @@ class LoadStats:
     def accumulated(cls, stats: list["LoadStats"]) -> "LoadStats":
         """Accumulate the stats across multiple LoadStats objects in a batch."""
         return cls(sum(s.load_time_seconds for s in stats))
+
+
+def compact_dataset(ds: xr.Dataset) -> xr.Dataset:
+    data = ds.copy()
+
+    var_groups = defaultdict(list)
+    for key in data.keys():
+        if "_lev_" in (k := str(key)):
+            base_name = k.split("_lev_")[0]
+            var_groups[base_name].append(k)
+
+    def _parse_level(x) -> float:
+        return float(x.split("_lev_")[1].replace("_", "."))
+
+    for base_var, vars_ in var_groups.items():
+        sorted_vars = sorted(vars_, key=_parse_level)
+        levels = [_parse_level(var) for var in sorted_vars]
+        if hasattr(data, "lev"):
+            levels = data.lev.values
+        da = xr.concat([data[var] for var in sorted_vars], dim="lev").assign_coords(
+            lev=("lev", levels)
+        )
+        data[base_var] = da
+        data = data.drop_vars(vars_)
+
+    return data
